Update token used in .pot file update step
The commit changes the token utilized in the .pot file update step in the GitHub workflows. Previously, the GIT_PUSH_TOKEN was used, but now it has been updated to use the GITHUB_TOKEN instead. This change is applied to the run command available within the 'Update .pot files' step.