Add computed field for tracking changed product fields
This commit introduces a 'changed_fields' computed field to the `cwa.import.product.change` model. It tracks differences in product attributes such as 'list_price' and 'ingredients' between the current state and the imported data. Additionally, a test case has been added to verify the functionality of this computed field.